# Performance tunings for awesome methods