The extensions to Array and a few other classes introduce a performance issue when awesome_print is loaded. This performance slowdown happens even when objects are not being printed. The main place this would be noticed in when taking the difference between two arrays.

This PR reduces the overhead of `Array#-` from **2.5x slower** to only **1.1x slower** and completely eliminates extraneous object creation.

The two main causes of the performance issue are
* The use of `define_method` to override methods. This requires the original method to be bound (allocating a Method object) and then called (allocating an Array of the args) on every method invocation.
* The use of unfrozen Strings where a Symbol can be used.
 
This [benchmark script] (https://gist.github.com/dgynn/0ee1ae8af279aa417f8aec5328ee6326) can be used to compare the performance before and after loading awesome_print and its extensions to Array and others. It tests the performance of...
```ruby
[1,2,3,4,5] - [3,4,5,6,7]
```

Here are the results of the benchmark script runs. "Before" means before requiring awesome_print and IPS means iterations per second. At least 1 object must be created in all cases which is the result Array.
```
# Initial results
Before IPS: 1363785.204974598
After IPS: 540003.9100471758
Speed difference: 2.5255098705775203
Before allocated: 40 bytes (1 objects)
After  allocated: 248 bytes (5 objects)
Memory difference: +208 bytes
Memory difference: +4 objects

# Results after converting Strings to Symbols for instance variables
Before IPS: 1331499.3192977062
After IPS: 565391.1058412988
Speed difference: 2.355005774836947
Before allocated: 40 bytes (1 objects)
After  allocated: 208 bytes (4 objects)
Memory difference: +168 bytes
Memory difference: +3 objects

# Results after directly overriding methods
Before IPS: 1441087.4199060071
After IPS: 1298607.328987272
Speed difference: 1.1097176088093152
Before allocated: 40 bytes (1 objects)
After  allocated: 40 bytes (1 objects)
Memory difference: +0 bytes
Memory difference: +0 objects
```