@gerrywastaken 2016-08-28
WOW! Very impressive work and a huge thank you. Are you able to add a license to your benchmark script? I'd like to include modifications of it in this conversation.