@dgynn 2016-08-28
Thanks! Sure, I've added a comment indicating that the gist is MIT licensed.