@gerrywastaken 2016-08-28
Firstly, for posterity I'll repost the original here (just as a backup)

Coppied from: https://gist.github.com/dgynn/0ee1ae8af279aa417f8aec5328ee6326
```ruby
#!/usr/bin/env ruby
# This script is licensed under the MIT license (http://www.opensource.org/licenses/mit-license.php)
# Feel free to copy, modify, reuse as you would like without limitation.
# 
# Save this file to the root directory where the awesome_print code is checked out.
#
# This script profiles the impact of awesome_print on taking difference of two arrays
#   [1,2,3,4,5] - [3,4,5,6,7]
# The script measures memory allocations and iterations/sec before and after
#   awesome_print has been loaded to demonstrate the impact
#
# This script can be run against a local version of Awesome Print or the latest release for comparison
# Run at least once with --install-gems to ensure all required Gems are installed
#   ruby ./benchmark.rb --install-gems
#
# To run benchmark against local code
#   ruby ./benchmark.rb --local
# To run benchmark against latest release
#   ruby ./benchmark.rb
#
# Use --help to see the other options

begin
  require 'bundler/inline'
  require 'optparse'
rescue LoadError => e
  $stderr.puts 'Bundler version 1.10 or later is required. Please update your Bundler'
  raise e
end

options = {:install_gems => false, :local => false}

ARGV.options do |opts|
  opts.banner = "Usage: benchmark.rb [options]"
  opts.on("--install-gems", "Bundle install the gems. Run with this flag once.") do |install|
    options[:install_gems] = install
  end
  opts.on("--local", "Run against local code rather than installed gem") do |local|
    options[:local] = local
  end

  opts.parse!
end

gemfile(options[:install_gems]) do
  source 'https://rubygems.org'

  if options[:local]
    gem 'awesome_print', :path => '.', :require=> false
  else
    gem 'awesome_print', :require=> false
  end

  gem 'memory_profiler'
  gem 'benchmark-ips', '>= 2.7.2'

end

ARRAY1 = [1,2,3,4,5]
ARRAY2 = [3,4,5,6,7]

def do_array_difference(arr1, arr2)
  arr1 - arr2
end

puts 'Before requiring awesome_print'

# Execute 1 time to initialize
do_array_difference(ARRAY1, ARRAY2)

puts 'Profiling array difference memory allocations'
result1 = MemoryProfiler.report do
  do_array_difference(ARRAY1, ARRAY2)
end #.pretty_print

puts 'Profiling array difference speed'
report1 = Benchmark.ips do |x|
  x.report('array difference w/o awesome_print loaded') { do_array_difference(ARRAY1, ARRAY2) }
end

require "awesome_print"
puts "\n\nAfter requiring awesome_print\n\n"

# Execute 1 time to initialize
do_array_difference(ARRAY1, ARRAY2)

puts 'Profiling array difference memory allocations'
result2 = MemoryProfiler.report do
  do_array_difference(ARRAY1, ARRAY2)
end #.pretty_print

report2 = Benchmark.ips do |x|
  x.report('array with awesome_print loaded') { do_array_difference(ARRAY1, ARRAY2) }
end

puts "\n\nSummary"
puts "Before IPS: #{report1.entries.first.ips}"
puts "After IPS: #{report2.entries.first.ips}"
puts "Speed difference: #{report1.entries.first.ips / report2.entries.first.ips }"

puts "Before allocated: #{result1.total_allocated_memsize} bytes (#{result1.total_allocated} objects)"
puts "After  allocated: #{result2.total_allocated_memsize} bytes (#{result2.total_allocated} objects)"
puts "Memory difference: +#{result2.total_allocated_memsize - result1.total_allocated_memsize } bytes"
puts "Memory difference: +#{result2.total_allocated - result1.total_allocated } objects"

# Uncomment to display
# puts 'Final memory_profiler report'
# result2.pretty_print
```