@gerrywastaken 2016-08-28
@dgynn I get very similar results to you when running your script, however the results are quite different when I simplify the script so that I can manually run it twice by just uncommenting or commenting the awesome_print require to see the difference.

```ruby
#!/usr/bin/env ruby
# This script is licensed under the MIT license (http://www.opensource.org/licenses/mit-license.php)
# Feel free to copy, modify, reuse as you would like without limitation.
require 'memory_profiler'
require 'benchmark/ips'

ARRAY1 = [1,2,[3,[4]],5]
ARRAY2 = [3,4,[5,[6]],7]

def do_array_difference(arr1, arr2)
  arr1 - arr2
end

# require "awesome_print"

puts 'Profiling memory'
# Execute 1 time to initialize
do_array_difference(ARRAY1, ARRAY2)
memory_results = MemoryProfiler.report do
  do_array_difference(ARRAY1, ARRAY2)
end

puts 'Profiling speed'
# Execute 1 time to initialize
do_array_difference(ARRAY1, ARRAY2)
speed_results = Benchmark.ips do |x|
  x.report('speed test') do
    do_array_difference(ARRAY1, ARRAY2)
  end
end

puts "\n\nSummary"
puts "Iterations per second: #{speed_results.entries.first.ips}"
puts "Allocated: #{memory_results.total_allocated_memsize} bytes (#{memory_results.total_allocated} objects)"
```

When I do this, the first run comes back with:

```
Profiling memory
Profiling speed
Warming up --------------------------------------
          speed test    23.781k i/100ms
Calculating -------------------------------------
          speed test    265.585k (± 5.3%) i/s -      1.332M in   5.029502s


Summary
Iterations per second: 265584.73857538856
Allocated: 200 bytes (1 objects)
```

Then if I uncomment the line with the require and run it again, I get:

```
Profiling memory
Profiling speed
Warming up --------------------------------------
          speed test    18.240k i/100ms
Calculating -------------------------------------
          speed test    203.718k (± 3.5%) i/s -      1.021M in   5.020190s


Summary
Iterations per second: 203717.57090932174
Allocated: 384 bytes (5 objects)
```

As you can see there is still a difference in speed, but not quite as dramatic as what you are seeing. I ran the script several times with each variation to make sure that I wasn't just seeing an outlier. I'm wondering if you can think of an obvious reason for this?