@dgynn 2016-08-28
I think the difference in relative timings is because the new test case compares shorter but more complex arrays.

Since we are really only trying to eliminate the **overhead** of awesome_print on any Array differencing there is a baseline performance of differencing arrays of different lengths. If I modify your script script so that both ARRAY1 and ARRAY2 are `[]` I see a difference of 5.671M vs 815.3.K iterations per second (~7x faster).

So the script timings are highly influenced by the size and characteristics of the arrays being differenced.