@gerrywastaken 2016-08-28
Oh wow, I completely forgot I had modified the array's structure and could not spot it when checking. Thanks for pointing that out.

My results after applying your PR are very similar:

```
~/source/awesome_print(master ✗) ./benchmark.rb --local
Before requiring awesome_print
Profiling array difference memory allocations
Profiling array difference speed
Warming up --------------------------------------
array difference w/o awesome_print loaded
                        73.770k i/100ms
Calculating -------------------------------------
array difference w/o awesome_print loaded
                        990.558k (± 5.2%) i/s -      4.943M in   5.004539s


After requiring awesome_print

Profiling array difference memory allocations
Warming up --------------------------------------
array with awesome_print loaded
                        39.017k i/100ms
Calculating -------------------------------------
array with awesome_print loaded
                        457.209k (± 4.5%) i/s -      2.302M in   5.045666s


Summary
Before IPS: 990557.8863353004
After IPS: 457209.05230610137
Speed difference: 2.166531658415464
Before allocated: 40 bytes (1 objects)
After  allocated: 224 bytes (5 objects)
Memory difference: +184 bytes
Memory difference: +4 objects
~/source/awesome_print(master ✗) git checkout pull/264; ./benchmark.rb --local
Switched to branch 'pull/264'
Your branch is up-to-date with 'upstream/pull/264'.
Before requiring awesome_print
Profiling array difference memory allocations
Profiling array difference speed
Warming up --------------------------------------
array difference w/o awesome_print loaded
                        74.114k i/100ms
Calculating -------------------------------------
array difference w/o awesome_print loaded
                        995.915k (± 5.2%) i/s -      4.966M in   4.999995s


After requiring awesome_print

Profiling array difference memory allocations
Warming up --------------------------------------
array with awesome_print loaded
                        69.530k i/100ms
Calculating -------------------------------------
array with awesome_print loaded
                        915.550k (± 5.0%) i/s -      4.589M in   5.025661s


Summary
Before IPS: 995914.8809517147
After IPS: 915550.0754901534
Speed difference: 1.0877776187376063
Before allocated: 40 bytes (1 objects)
After  allocated: 40 bytes (1 objects)
Memory difference: +0 bytes
Memory difference: +0 objects
~/source/awesome_print(pull/264 ✗)
```

Again, fantastic work!

Also thanks @kikonen for originally pointing to the string allocation issue and mentioning that there were likely other memory allocation issues associated with the code.

Merging